Working on the build system part 2: Standardize 'component' structure so boilerplate works, split mfmtest from mfmsim and build both
