+++ Makefile
@@ -1,36 +1,16 @@
-CC := g++ 
-CFLAGS += -ansi -pedantic -Wall -Werror -O99
-ALLDEP := Makefile
-SRCDIR := src
-BUILDDIR := build
-OUTPUTDIR := bin
-TESTTARGET:= $(OUTPUTDIR)/mfm_test
-TARGET := $(OUTPUTDIR)/mfm
+# For now, we'll assume we're building the simulator only
+PLATFORMS=sim
 
-SIMMAIN := $(SRCDIR)/drivers/MFMSim.cpp
+.PHONY:	$(PLATFORMS) all clean realclean
 
-SRCEXT := cpp
-OBJEXT := o
-HDRPAT := -name *.h -o -name *.tcc
-SOURCES := $(shell find $(SRCDIR) -type f -name *.$(SRCEXT))
-HEADERS := $(shell find $(SRCDIR) -type f $(HDRPAT))
-ALLDEP += $(HEADERS)
-OBJECTS := $(shell find $(BUILDDIR) -name *.$(OBJEXT))
+sim:	PLATFORMS:=sim
 
-LIB := -L lib -lm -lSDL -L build
-INC := -I src/core/include -I src/gui/include -I src/sim/include
+all:	$(PLATFORMS)
 
-$(TARGET): dependencies
-	$(CC) $(CFLAGS) -o $(TARGET) $(SIMMAIN) $(LIB) $(OBJECTS) $(INC)
+clean:  $(PLATFORMS)
 
-dependencies:
-	@mkdir -p $(OUTPUTDIR)
-	@make -C src/core
-	@make -C src/sim
-	@make -C src/gui
+realclean:  $(PLATFORMS)
 
-clean:
-	@echo "Cleaning..."
-	@echo "$(RM) -r $(BUILDDIR) $(TARGET) $(OUTPUTDIR)"; $(RM) -r $(BUILDDIR) $(TARGET) $(OUTPUTDIR)
-
-.PHONY: clean
+# Pass each entry in PLATFORMS down as a target
+$(PLATFORMS):
+	export TARGET=$@;$(MAKE) -C src $(MAKECMDGOALS)
