+++ src/drivers/mfmsim/include/main.h
@@ -1,0 +1,22 @@
+#ifndef MAIN_H
+#define MAIN_H
+
+#include <stdio.h>
+#include "SDL/SDL.h"
+#include "assert.h"
+#include "eventwindow.h"
+#include "grid.h"
+#include "gridpanel.h"
+#include "gridrenderer.h"
+#include "itype.h"
+#include "keyboard.h"
+#include "manhattandir.h"
+#include "mouse.h"
+#include "p1atom.h"
+#include "point.h"
+#include "drawing.h"
+#include "tilerenderer.h"
+#include "time.h"
+
+#endif /* MAIN_H */
+
