+++ src/drivers/mfmtest/include/main.h
@@ -1,0 +1,6 @@
+#ifndef MAIN_H
+#define MAIN_H
+
+#include "tests.h"
+
+#endif  /* MAIN_H */
